backend/requirements/common.txt
Django==4.2.1
python-dotenv==1.0.0
djangorestframework==3.14.0
django-filter==23.2
psycopg==3.1.9
psycopg2-binary==2.9.6
django-allauth==0.54.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
django-rest-auth==0.9.5
django-extensions==3.2.1